Fix error handling on bus acquisition
Error codes are just enumeration values, and thus they can overlap
across different error types. This is why GError has a domain field, as
well as a `g_error_matches()` function. Instead of just checking the
error code, we should do a full check on the error domain as well.

Additionally, since both Collector and Provider instances should be safe
to use even in the case of a failure to connect to the session bus, we
should not use a warning.